mode notifica servidores + tests arreglados

+++ src/server/connection_handler/client_handler/logic/mode_logic.rs
@@ -1,16 +1,19 @@
 use std::io;
 
-use crate::server::{
-    connection::Connection,
-    connection_handler::{
-        client_handler::ClientHandler,
-        mode_requests::{ChannelModeRequest, UserModeRequest},
+use crate::{
+    macros::ok_or_return,
+    server::{
+        connection::Connection,
+        connection_handler::{
+            client_handler::ClientHandler,
+            mode_requests::{ChannelModeRequest, UserModeRequest},
+        },
+        consts::{
+            commands::MODE_COMMAND,
+            modes::{ChannelFlag, UserFlag},
+        },
+        responses::ErrorReply,
     },
-    consts::{
-        commands::MODE_COMMAND,
-        modes::{ChannelFlag, UserFlag},
-    },
-    responses::ErrorReply,
 };
 
 impl<C: Connection> ClientHandler<C> {
@@ -49,18 +52,12 @@ impl<C: Connection> ClientHandler<C> {
             }
             ChannelModeRequest::GetBanmasks => self.get_banmasks_request(channel),
         }
-
-        // let sender = &self.nickname;
-        // let target = channel;
-        // let request = request.to_string();
-
-        // let notification = Notification::mode(sender, target, &request);
-        // self.send_message_to_local_clients_on_channel(&notification, channel);
-        // self.send_message_to_all_servers(&notification);
     }
 
     fn add_banmask_request(&mut self, channel: &str, banmask: String) -> io::Result<()> {
         self.database.add_channel_banmask(channel, &banmask);
+        let request = ChannelModeRequest::AddBanmask(banmask);
+        self.send_channel_mode_request_notification(channel, request);
 
         Ok(())
     }
@@ -76,6 +73,9 @@ impl<C: Connection> ClientHandler<C> {
 
         self.database.add_channop(channel, &operator);
 
+        let request = ChannelModeRequest::AddOperator(operator);
+        self.send_channel_mode_request_notification(channel, request);
+
         Ok(())
     }
 
@@ -85,10 +85,20 @@ impl<C: Connection> ClientHandler<C> {
         }
         self.database.add_speaker(channel, &speaker);
 
+        let request = ChannelModeRequest::AddSpeaker(speaker);
+        self.send_channel_mode_request_notification(channel, request);
+
         Ok(())
     }
     fn remove_banmask_request(&mut self, channel: &str, banmask: String) -> io::Result<()> {
+        let banmasks = ok_or_return!(self.database.get_channel_banmask(channel), Ok(()));
+        if !banmasks.contains(&banmask) {
+            return Ok(());
+        }
         self.database.remove_channel_banmask(channel, &banmask);
+
+        let request = ChannelModeRequest::RemoveBanmask(banmask);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn remove_operator_request(&mut self, channel: &str, operator: String) -> io::Result<()> {
@@ -96,6 +106,9 @@ impl<C: Connection> ClientHandler<C> {
             return self.stream.send(&error);
         }
         self.database.remove_channop(channel, &operator);
+
+        let request = ChannelModeRequest::RemoveOperator(operator);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn remove_speaker_request(&mut self, channel: &str, speaker: String) -> io::Result<()> {
@@ -103,38 +116,57 @@ impl<C: Connection> ClientHandler<C> {
             return self.stream.send(&error);
         }
         self.database.remove_speaker(channel, &speaker);
+
+        let request = ChannelModeRequest::RemoveSpeaker(speaker);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn set_channel_flag_request(&mut self, channel: &str, flag: ChannelFlag) -> io::Result<()> {
-        self.database.set_channel_mode(channel, flag);
+        self.database.set_channel_mode(channel, flag.clone());
 
+        let request = ChannelModeRequest::SetFlag(flag);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn set_key_request(&mut self, channel: &str, key: String) -> io::Result<()> {
         if let Err(error) = self.assert_can_set_key(channel) {
             return self.stream.send(&error);
         }
-        self.database.set_channel_key(channel, Some(key));
+        self.database.set_channel_key(channel, Some(key.clone()));
+
+        let request = ChannelModeRequest::SetKey(key);
+        self.send_channel_mode_request_notification(channel, request);
 
         Ok(())
     }
     fn set_limit_request(&mut self, channel: &str, limit: usize) -> io::Result<()> {
         self.database.set_channel_limit(channel, Some(limit));
 
+        let request = ChannelModeRequest::SetLimit(limit);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn unset_limit_request(&mut self, channel: &str) -> io::Result<()> {
         self.database.set_channel_limit(channel, None);
+
+        let request = ChannelModeRequest::UnsetLimit();
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn unset_key_request(&mut self, channel: &str) -> io::Result<()> {
         self.database.set_channel_key(channel, None);
+
+        let request = ChannelModeRequest::UnsetKey();
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn unset_channel_flag_request(&mut self, channel: &str, flag: ChannelFlag) -> io::Result<()> {
-        if self.database.channel_has_mode(channel, &flag) {
-            self.database.unset_channel_mode(channel, flag)
+        if !self.database.channel_has_mode(channel, &flag) {
+            return Ok(());
         }
+        self.database.unset_channel_mode(channel, flag.clone());
+        let request = ChannelModeRequest::UnsetFlag(flag);
+        self.send_channel_mode_request_notification(channel, request);
         Ok(())
     }
     fn unknown_channel_mode_request(&mut self, character: char) -> io::Result<()> {
@@ -165,12 +197,28 @@ impl<C: Connection> ClientHandler<C> {
     }
 
     fn set_user_flag_request(&mut self, flag: UserFlag, user: &str) -> io::Result<()> {
-        self.database.set_user_mode(user, flag);
+        if flag == UserFlag::Operator {
+            return Ok(());
+        }
+
+        self.database.set_user_mode(user, flag.clone());
+
+        let request = UserModeRequest::SetFlag(flag);
+        self.send_user_mode_request_notification(request, user);
 
         Ok(())
     }
+
     fn unset_user_flag_request(&mut self, flag: UserFlag, user: &str) -> io::Result<()> {
-        self.database.unset_user_mode(user, flag);
+        let info = ok_or_return!(self.database.get_client_info(user), Ok(()));
+
+        if !info.flags.contains_key(&flag) {
+            return Ok(());
+        }
+        self.database.unset_user_mode(user, flag.clone());
+
+        let request = UserModeRequest::UnsetFlag(flag);
+        self.send_user_mode_request_notification(request, user);
 
         Ok(())
     }
